Исправлен класс с тестами
