+++ src/Epic.java
@@ -11,6 +11,9 @@ public class Epic extends Task {
     }
 
     public void addSubtask(int subtaskId) {
+        if (subtaskId == this.getId()) {
+            throw new IllegalArgumentException("Epic не может быть своим сабтаском");
+        }
         subtaskIds.add(subtaskId);
     }
