+++ src/InMemoryHistoryManager.java
@@ -22,9 +22,10 @@ public class InMemoryHistoryManager implements HistoryManager {
             return;
         } else {
             remove(task.getId());
-            LinkLast(task);
+            linkLast(task);
         }
-        if (history.size() > 10) {
+        if (history.size() > 10
+        ) {
             Integer firstKey = history.keySet().iterator().next();
             history.remove(firstKey);
         }
@@ -41,7 +42,7 @@ public class InMemoryHistoryManager implements HistoryManager {
         return history;
     }
 
-    public void LinkLast(Task task) {
+    public void linkLast(Task task) {
         final Node oldTail = tail;
         final Node newNode = new Node(oldTail, task, null);
         tail = newNode;
